firebase app initialization moved to utils file
